Add user msg, email deletetion first time -> to trash, second time ->delete forever.

+++ src/pages/EmailIndex.jsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react"
-import { useLocation, useNavigate, Outlet, useParams } from 'react-router-dom'
+import { useLocation, useNavigate, Outlet, useParams, useSearchParams } from 'react-router-dom'
 
 
 import { emailService } from "../services/EmailService"
@@ -9,26 +9,37 @@ import { MenuBar } from '../cmps/MenuBar.jsx';
 import { EmailList } from "../cmps/EmailList"
 import { AppFooter } from '../cmps/AppFooter.jsx';
 
+import { eventBusService, showErrorMsg, showSuccessMsg } from "../services/event-bus.service"
+
+
 
 export function EmailIndex() {
     const params = useParams()
     const navigate = useNavigate()
 
+    const [searchParams, setSearchParams] = useSearchParams()
     const [emails, setEmails] = useState(null)
-    const [filterBy, setFilterBy] = useState(emailService.getDefaultFilter(params.folder))
+    const [filterBy, setFilterBy] = useState(emailService.getFilterFromParams(searchParams))
+    const [unreadCount, setUnreadCount] = useState(null)
+
 
     const { folder, txt, isRead } = filterBy
 
     useEffect(() => {
-        // console.log("params", params);
-        // if (!params.folder)
-        //     navigate('inbox')
+        getUnreadCountFromService()
+    })
 
-        console.log("filterby:", filterBy);
+    useEffect(() => {
+        setSearchParams(filterBy)
         loadEmails()
     }, [filterBy])
     /* use effect listening to changes on filterBy that occure in appHeader and menuBar*/
 
+
+    async function getUnreadCountFromService() {
+        const count = await emailService.getUnreadCount()
+        setUnreadCount(count)
+    }
 
     function onSetFilter(fieldsToUpdate) {
         setFilterBy((prevFilterBy) => ({ ...prevFilterBy, ...fieldsToUpdate }))
@@ -47,33 +58,61 @@ export function EmailIndex() {
 
     async function onRemoveEmail(emailId) {
         try {
-            await emailService.remove(emailId)
+            let email = await emailService.getById(emailId)
+            if (!email.removedAt) {
+                email.removedAt = new Date();
+                await emailService.save(email)
+                showSuccessMsg('Conversation moved to Trash.')
+            } else {
+                await emailService.remove(emailId)
+                showSuccessMsg('Conversation deleted forever.')
+            }
+
+
+
             setEmails((prevEmails) => {
                 return prevEmails.filter(email => email.id !== emailId)
             })
+
+
+
         } catch (err) {
             console.log('Error in onRemoveEmail', err)
+            showErrorMsg('Could not remove email')
+
         }
     }
 
     async function onSetIsRead(isRead, emailId) {
 
-        console.log("in onSetIsRead");
         try {
             let email = await emailService.getById(emailId)
             email.isRead = !isRead
             await emailService.save(email)
             setEmails((prevEmails) => {
-                return prevEmails.map(email => {
-                    if (email.id == emailId)
+                return prevEmails.filter(email => {
+                    if (email.id == emailId) {
                         email.isRead = !isRead
-                    return email
+                        updateUnreadCount(email)
+                        if (email.isRead != searchParams.get('isRead') && searchParams.get('isRead') != 'null') {
+                            return false
+                        }
+                    }
+                    return true
                 })
             })
         } catch (err) {
             console.log('Error in onSetIsRead', err)
         }
     }
+
+    function updateUnreadCount(email) {
+        if (!email.isRead && (email.to === emailService.loggedinUser.email && !email.removedAt))
+            setUnreadCount(prevCount => prevCount++)
+
+    }
+
+
 
     async function onStaring(emailId) {
         try {
@@ -97,7 +136,7 @@ export function EmailIndex() {
     return (
         <section className='main-app' >
             <AppHeader filterBy={{ txt, isRead }} onSetFilter={onSetFilter} />
-            <MenuBar filterBy={{ folder }} onSetFilter={onSetFilter} />
+            <MenuBar filterBy={{ folder }} onSetFilter={onSetFilter} unreadCount={unreadCount} />
             <div className="main-content">
 
                 {!params.emailId && <EmailList emails={emails} onRemoveEmail={onRemoveEmail} onSetIsRead={onSetIsRead} onStaring={onStaring} />}
